+++ relay/Cargo.toml
@@ -13,4 +13,6 @@ tracing-subscriber = "0.3.17"
 tracing = "0.1.39"
 webrtc = "0.9.0"
 anyhow = "1"
+serde = "1"
+serde_json = "1"
 [dev-dependencies]
